scripts/requirements.txt
python-dotenv==1.0.1
langchain-community==0.3.14
langchain-text-splitters==0.3.4
upstash-vector==0.6.0
pypdf==5.1.0
httpx==0.28.1
